+++ config/config.go
@@ -38,6 +38,9 @@ var bkmkPath string
 
 var DownloadsDir string
 
+// Command for opening HTTP(S) URLs in the browser, from "a-general.http" in config.
+var HttpCommand []string
+
 //nolint:golint,goerr113
 func Init() error {
 
@@ -236,6 +239,15 @@ func Init() error {
 	if viper.GetBool("a-general.color") {
 		cview.Styles.PrimitiveBackgroundColor = GetColor("bg")
 	} // Otherwise it's black by default
+
+	// Parse HTTP command
+	HttpCommand = viper.GetStringSlice("a-general.http")
+	if len(HttpCommand) == 0 {
+		// Not a string array, interpret as a string instead
+		// Split on spaces to maintain compatibility with old versions
+		// The new better way to is to just define a string array in config
+		HttpCommand = strings.Fields(viper.GetString("a-general.http"))
+	}
 
 	return nil
 }